+++ modules/device/include/ifm3d/device/o3r.h
@@ -220,6 +220,8 @@ namespace ifm3d
      */
     void FromJSON(const json& j) override;
 
+    void DownloadServiceReport();
+
   private:
     class Impl;
     std::unique_ptr<Impl> pImpl;